implement dead time

+++ device_info.go
@@ -6,9 +6,9 @@ type DeviceInfo struct {
 	Latitude        float64
 	Longitude       float64
 	AlertThreshold  float64
-	AlertDuration   float64
-	AlertCount      int
-	AlertDeadtime   float64
+	AlertDuration   int64
+	AlertCount      int64
+	AlertDeadtime   int64
 	AlertPhone      string
 	AlertActive     bool
 	TurnOnTime      int
